+++ Bank.sln
@@ -13,8 +13,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Models", "Models\Models.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PracticeWithTypes", "PracticeWithTypes\PracticeWithTypes.csproj", "{91C2161A-505D-4D59-8A30-572B93CFF5C7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BankService", "BankService\BankService.csproj", "{001380B2-D3D7-4169-9379-64DED7EE6431}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services", "Services\Services.csproj", "{68516DFA-9663-43F0-8F5D-D4F21F7CECE3}"
 	ProjectSection(ProjectDependencies) = postProject
 		{3293480B-3D47-4DC8-BEC3-12EA1EB0EE9D} = {3293480B-3D47-4DC8-BEC3-12EA1EB0EE9D}
@@ -40,10 +38,6 @@ Global
 		{91C2161A-505D-4D59-8A30-572B93CFF5C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{91C2161A-505D-4D59-8A30-572B93CFF5C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{91C2161A-505D-4D59-8A30-572B93CFF5C7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{001380B2-D3D7-4169-9379-64DED7EE6431}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{001380B2-D3D7-4169-9379-64DED7EE6431}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{001380B2-D3D7-4169-9379-64DED7EE6431}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{001380B2-D3D7-4169-9379-64DED7EE6431}.Release|Any CPU.Build.0 = Release|Any CPU
 		{68516DFA-9663-43F0-8F5D-D4F21F7CECE3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{68516DFA-9663-43F0-8F5D-D4F21F7CECE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{68516DFA-9663-43F0-8F5D-D4F21F7CECE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -59,7 +53,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{3293480B-3D47-4DC8-BEC3-12EA1EB0EE9D} = {2DAB60D7-5A6B-4CEC-A89D-2D0EA563B334}
 		{91C2161A-505D-4D59-8A30-572B93CFF5C7} = {BCDD88E8-992E-4798-A26E-12FB1F55E5A5}
-		{001380B2-D3D7-4169-9379-64DED7EE6431} = {2DAB60D7-5A6B-4CEC-A89D-2D0EA563B334}
 		{68516DFA-9663-43F0-8F5D-D4F21F7CECE3} = {2DAB60D7-5A6B-4CEC-A89D-2D0EA563B334}
 		{67862A47-FF20-407D-AA7D-FC6766A445EC} = {BCDD88E8-992E-4798-A26E-12FB1F55E5A5}
 	EndGlobalSection
