+++ PracticeWithTypes/PracticeWithTypes.csproj
@@ -8,7 +8,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\BankService\BankService.csproj" />
+      <ProjectReference Include="..\Services\Services.csproj" />
       <ProjectReference Include="..\Models\Models.csproj" />
     </ItemGroup>
